+++ R/midnight-package.R
@@ -3,5 +3,7 @@
 
 ## usethis namespace: start
 #' @importFrom rlang '%||%'
+#' @importFrom rlang .data
+#' @importFrom parsnip new_model_spec
 ## usethis namespace: end
 NULL